Edit mode multi-select, hOCR refactor, page-preserving navigation, and crop info in RecogniseMode (#82)
Edit mode multi-select, hOCR refactor, page-preserving navigation, and crop info in RecogniseMode

Summary

* Adds multi-item selection in Edit Mode (Shift+Click to toggle items, highlights in both viewport and outline)
* Introduces new modifier-key scheme: Cmd for split/join, Cmd+Shift for add, Cmd+Alt+Shift for remove
* Batch keyboard actions on all selected items: Delete/Backspace removes, number keys 0–7 change block type, ArrowUp/Down moves
* Creates a useHocr composable and context pattern; centralises hOCR loading in PageWorkspace and shares it across Edit and Recognise modes
* Adds provideHocrContext() to IngestorView, InspectMode, and CropMode to prevent injection errors
* Preserves the current page stem in the URL when switching between any mode (Inspect, Crop, Edit, Recognise, Ingestor)
* Displays crop region values (Top, Left, Right, Bottom) and enables the crop overlay in RecogniseMode
* Fixes image block persistence: ocr_photo class and <img> tags now round-trip correctly through the Rust hOCR parser
* Hides the "Block type" radio group when adding CAREAs (only relevant for block-level additions)
* Adds an application-wide error toast for endpoint and API failures
* Fixes move up/down ordering bugs in the hOCR handler, an accumulator issue in CropMode, and hOCR context injection errors
* Wraps HocrBbox in a dedicated struct with utility methods; implements add carea/block with automatic removal of overlapping careas
* Removes obsolete Jinja/minijinja HTML templates
* Expands planning documentation with a roadmap for Folios, OmniGraph, Codex, and a sketch of the future multi-page Assist section

Test plan

* [ ] In Edit Mode, click an item — only that item should be highlighted in both the viewport and the outline
* [ ] Shift+click additional items — each should be added to the selection; all highlighted simultaneously
* [ ] With multiple items selected, press Delete — all selected items should be removed in one action
* [ ] With multiple block items selected, press a number key (0–7) — all selected blocks should change to that type
* [ ] With multiple items selected, use the move-up/down keyboard shortcut — all move together in the correct order
* [ ] Hold Cmd (Mac) / Ctrl (Win) and click between two items — split/join should be triggered
* [ ] Hold Cmd+Shift and drag — a new item should be created at the drawn bounding box
* [ ] Hold Cmd+Alt+Shift and click an item — it should be removed
* [ ] Select "CAREA" tool in Add mode — the "Block type" radio group should NOT appear; switch to "Block" tool — it should appear
* [ ] Add an image block, navigate to another page, navigate back — the image block should still be present
* [ ] Navigate to a specific page in Inspect mode, then click Crop in the nav bar — the same page stem should appear in the URL and the same page should be displayed
* [ ] Navigate from Folios to the Ingestor — the page parameter should carry over to the Ingestor URL
* [ ] Open RecogniseMode — the green/red crop overlay should be visible on the page preview, and Top/Left/Right/Bottom crop values should appear in the tools panel
* [ ] Trigger a backend error (e.g., a bad API call) — an error toast should appear
* [ ] Navigate to Ingestor, InspectMode, and CropMode — no hOCR context errors should appear in the browser console

+++ Apps/OmniGraph/src/routes/projects/forms.rs
@@ -77,7 +77,7 @@ pub struct IngestQuery {
     pub before: Option<usize>,
 }
 
-#[derive(Deserialize, Serialize)]
+#[derive(Deserialize, Serialize, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
 pub enum AddBlockType {
     Text,
@@ -93,7 +93,7 @@ pub struct AddRequest {
     pub block_type: Option<AddBlockType>,
     pub text: Option<String>,
     pub erase_underneath: Option<bool>,
-    pub erase_overlap: Option<usize>
+    pub erase_overlap: Option<u8>
 }
 
 #[derive(Deserialize, Serialize)]
